backend/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
opencv-python-headless
Pillow
numpy
scikit-image
albumentations
jinja2
weasyprint
python-multipart